requirements.txt
anyio==4.9.0
boto3==1.37.25
boto3-stubs==1.37.25
botocore==1.37.25
botocore-stubs==1.37.25
certifi==2025.1.31
charset-normalizer==3.4.1
dynamodb-json==1.4.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
mypy-boto3-cloudformation==1.37.22
mypy-boto3-dynamodb==1.37.12
mypy-boto3-ec2==1.37.24
mypy-boto3-lambda==1.37.16
mypy-boto3-rds==1.37.21
mypy-boto3-s3==1.37.24
mypy-boto3-sqs==1.37.0
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
Telethon==1.39.0
types-awscrt==0.24.2
types-s3transfer==0.11.4
typing_extensions==4.13.0
urllib3==2.3.0
python-binance==1.0.29
